requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.123
certifi==2024.6.2
Flask==3.0.3
Flask-Cors==4.0.1
google-api-python-client==2.132.0
google-cloud-vision==3.7.4
mediapipe==0.10.14
moviepy==1.0.3
numpy==1.26.4
openai==1.33.0
gunicorn==23.0.0
opencv-python-headless==4.10.0.84
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
python-dotenv==1.0.1
pytube==15.0.0
requests==2.32.3
scenedetect==0.6.3
selenium==4.21.0
serpapi==0.1.5
torch==2.4.0
ultralytics==8.2.74
ultralytics-thop==2.0.0
urllib3==2.2.1
webdriver-manager==4.0.2